Tweak keyboard shortcuts to local preferences.
I prefer to use the home keys for switching, and cursor movements.

Mod-a to switch to the browser, Mod-d to switch to the terminal, Mod-f
to switch to the hexchat IRC terminal.

Floating (Mod-f) is moved to Mod-o (mnemonic: "orbiting") to prevent a
clash. Similarlly, Mod-d is switched to Mod-u (mnemonic:"unincrement")

Mod-h,j,k,l are switched to Mod-shift-h,j,k,l as I use the former for
cursor keys via a custom xkb layout.